Fix CodeQL warnings: unsigned subtraction and integer overflow
Alert #2 (cpp/unsigned-difference-expression-compared-zero):
src/common/utils/var_buf/xqc_var_buf.c:133 compared unsigned subtraction
to zero, which wraps on underflow instead of going negative. Changed to
compare the operands directly before subtracting.

Alert #1 (cpp/integer-multiplication-cast-to-long):
src/transport/fec_schemes/xqc_galois_calculation.c:81 multiplied two ints
and passed the result to memset's size_t parameter, risking overflow in
the int space before the widening cast. Cast one operand to size_t first
so the multiplication happens in size_t.

+++ src/transport/fec_schemes/xqc_galois_calculation.c
@@ -78,7 +78,7 @@ xqc_submatrix(int row_min, int row_max,
     int col_max_sub, int col_max_matrix,
     unsigned char *submatrix, unsigned char *matrix)
 {
-    xqc_memset(submatrix, 0, col_max_sub * row_max);
+    xqc_memset(submatrix, 0, (size_t)col_max_sub * row_max);
     for (int row_i = row_min; row_i < row_max; row_i++) {
         for (int col_i = col_min; col_i < col_max; col_i++) {
             *(submatrix + (row_i - row_min) * col_max_sub + col_i - col_min) = *(matrix + row_i * col_max_matrix + col_i);
